feat: add client cvar to disable screenshake
