+++ Content.Shared/Gravity/SharedGravitySystem.Shake.cs
@@ -9,7 +9,7 @@ public abstract partial class SharedGravitySystem
 {
     // ES START
     [Dependency] private readonly SharedGameTicker _ticker = default!;
-    [Dependency] private readonly ESScreenshakeSystem _shake = default!;
+    [Dependency] private readonly SharedESScreenshakeSystem _shake = default!;
     // ES END
 
     protected const float GravityKick = 100.0f;
